side-project/gpt/gpt-api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["dllchun <wingchuncheung0108@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.4"
langchain-openai = "^0.0.5"
pymongo = "^4.6.1"
certifi = "^2023.11.17"
datetime = "^5.4"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
